fix: created location&time component
