+++ src/views/components/injected/CourseCatalogInjectedPopup/HeadingAndActions.tsx
@@ -20,13 +20,9 @@ import OpenNewIcon from '~icons/material-symbols/open-in-new';
 import Remove from '~icons/material-symbols/remove';
 import Reviews from '~icons/material-symbols/reviews';
 
-const { openNewTab, addCourse, removeCourse, openCESPage } = background;
+import LocationAndTime from './LocationAndTime';
 
-const InstructionMode = {
-    Online: 'Online',
-    InPerson: 'inPerson',
-    Hybrid: 'hybrid',
-} as const;
+const { openNewTab, addCourse, removeCourse, openCESPage } = background;
 
 interface HeadingAndActionProps {
     /* The course to display */
@@ -55,7 +51,6 @@ const capitalizeString = (str: string) => str.charAt(0).toUpperCase() + str.slic
  * @returns {JSX.Element} The rendered component.
  */
 export default function HeadingAndActions({ course, activeSchedule, onClose }: HeadingAndActionProps): JSX.Element {
-    console.log(course);
     const {
         courseName,
         department,
@@ -75,9 +70,6 @@ export default function HeadingAndActions({ course, activeSchedule, onClose }: H
         if (firstName === '') return capitalizeString(lastName);
         return `${capitalizeString(firstName)} ${capitalizeString(lastName)}`;
     };
-
-    const getBuildingUrl = (building: string) =>
-        `https://utdirect.utexas.edu/apps/campus/buildings/nlogon/maps/UTM/${building}`;
 
     const handleCopy = () => {
         navigator.clipboard.writeText(formattedUniqueId);
@@ -172,61 +164,7 @@ export default function HeadingAndActions({ course, activeSchedule, onClose }: H
                         ))}
                     </div>
                 </div>
-                <div className='mt-1 flex flex-col'>
-                    {Array.isArray(schedule.meetings) && schedule.meetings.length > 0
-                        ? schedule.meetings.map(meeting => {
-                              console.log(meeting);
-                              const daysString = meeting.getDaysString({ format: 'long', separator: 'long' });
-                              const timeString = meeting.getTimeString({ separator: ' to ', capitalize: false });
-
-                              let locationInfo: string | JSX.Element = '';
-                              if (meeting.location) {
-                                  locationInfo = (
-                                      <>
-                                          {'in '}
-                                          <Link
-                                              href={getBuildingUrl(meeting.location.building)}
-                                              className='link'
-                                              variant='h4'
-                                          >
-                                              {meeting.location.building} {meeting.location.room}
-                                          </Link>
-                                      </>
-                                  );
-                              } else if (instructionMode !== 'Online') {
-                                  locationInfo = '(No location has been provided)';
-                              }
-
-                              return (
-                                  <Text
-                                      key={
-                                          daysString +
-                                          timeString +
-                                          (meeting.location?.building ?? '') +
-                                          (meeting.location?.room ?? '')
-                                      }
-                                      variant='h4'
-                                      as='p'
-                                  >
-                                      {daysString} {timeString} {locationInfo}
-                                  </Text>
-                              );
-                          })
-                        : (() => {
-                              if (instructionMode !== 'Online') {
-                                  return (
-                                      <Text variant='h4' as='p'>
-                                          No time and location has been provided.
-                                      </Text>
-                                  );
-                              }
-                              return (
-                                  <Text variant='h4' as='p'>
-                                      No time has been provided.
-                                  </Text>
-                              );
-                          })()}
-                </div>
+                <LocationAndTime course={course} />
             </div>
             <div className='my-3 flex flex-wrap items-center gap-x-3.75 gap-y-2.5'>
                 <Button
